chore: deprecate the rest of selector-utils
- move current exported api to deprecation folder
- mark as deprecated on index exports
