+++ packages/core/test/deprecated/deprecated-selector-utils.spec.ts
@@ -1,5 +1,11 @@
 import { expect } from 'chai';
-import { matchSelectorTarget } from '@stylable/core';
+import {
+    matchSelectorTarget,
+    filterChunkNodesByType,
+    separateChunks,
+    parseSelector,
+    SelectorChunk,
+} from '@stylable/core';
 
 describe('deprecated/selector-utils', () => {
     describe('matchSelectorTarget', () => {
@@ -57,6 +63,156 @@ describe('deprecated/selector-utils', () => {
             expect(matchSelectorTarget('.x::y.x.z', '.x::y.z'), '2').to.equal(true);
             expect(matchSelectorTarget('.x::y.x.x.x::z.z', '.x::y'), '3').to.equal(false);
             expect(matchSelectorTarget('.x.x.x::y.z', '.x::y.z'), '4').to.equal(true);
+        });
+    });
+
+    describe('filterChunkNodesByType', () => {
+        it('should filter and return only selector nodes which match types specified in array', () => {
+            expect(
+                filterChunkNodesByType({ nodes: [{ name: '0', type: 'a' }], type: 'dont-care' }, [
+                    'a',
+                ])
+            ).to.eql([
+                {
+                    name: '0',
+                    type: 'a',
+                },
+            ]);
+            expect(
+                filterChunkNodesByType(
+                    {
+                        nodes: [
+                            { name: '0', type: 'a' },
+                            { name: '1', type: 'b' },
+                            { name: '2', type: 'c' },
+                        ],
+                        type: 'dont-care',
+                    },
+                    ['b', 'a']
+                )
+            ).to.eql([
+                { name: '0', type: 'a' },
+                { name: '1', type: 'b' },
+            ]);
+        });
+    });
+
+    describe('separateChunks', () => {
+        const seperateChunksTestVectors: Array<{
+            title: string;
+            selector: string;
+            expected: SelectorChunk[][];
+        }> = [
+            {
+                title: 'empty selector',
+                selector: '',
+                expected: [[{ type: 'selector', nodes: [] }]],
+            },
+            {
+                title: 'class in first chunk',
+                selector: '.x',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'x' }],
+                        },
+                    ],
+                ],
+            },
+            {
+                title: 'handle spacing',
+                selector: '.x .y',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'x' }],
+                        },
+                        {
+                            type: 'spacing',
+                            value: ' ',
+                            nodes: [{ type: 'class', name: 'y' }],
+                        },
+                    ],
+                ],
+            },
+            {
+                title: 'handle operator',
+                selector: '.x + .y',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'x' }],
+                        },
+                        {
+                            type: 'operator',
+                            operator: '+',
+                            nodes: [{ type: 'class', name: 'y' }],
+                        },
+                    ],
+                ],
+            },
+            {
+                title: 'handle multiple selector',
+                selector: '.x, .y',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'x' }],
+                        },
+                    ],
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'y' }],
+                        },
+                    ],
+                ],
+            },
+            {
+                title: 'handle chunks with several nodes',
+                selector: '.x, .y::z',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [{ type: 'class', name: 'x' }],
+                        },
+                    ],
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [
+                                { type: 'class', name: 'y' },
+                                { type: 'pseudo-element', name: 'z' },
+                            ],
+                        },
+                    ],
+                ],
+            },
+            {
+                title: 'handle 2 selectors',
+                selector: '.x.y',
+                expected: [
+                    [
+                        {
+                            type: 'selector',
+                            nodes: [
+                                { type: 'class', name: 'x' },
+                                { type: 'class', name: 'y' },
+                            ],
+                        },
+                    ],
+                ],
+            },
+        ];
+        seperateChunksTestVectors.forEach((test) => {
+            it(test.title, () => {
+                expect(separateChunks(parseSelector(test.selector))).to.eql(test.expected);
+            });
         });
     });
 });
